+++ unidec_src/UniDec/UD_dataproc.h
@@ -16,9 +16,9 @@
 #define DATAPROC_HEADER
 
 //Pool data
-int pool1d(double *oldmz, double *oldint, const int oldlen, const int mzbins)
+int pool1d(float *oldmz, float *oldint, const int oldlen, const int mzbins)
 {
-	double *newmz, *newint;
+	float *newmz, *newint;
 	int newlen = oldlen/mzbins;
 
 	if (oldlen > newlen*mzbins)
@@ -27,15 +27,15 @@ int pool1d(double *oldmz, double *oldint, const int oldlen, const int mzbins)
 	}
 	
 	//declare memory for new arrays
-	newmz = calloc(newlen, sizeof(double));
-	newint = calloc(newlen, sizeof(double));
+	newmz = calloc(newlen, sizeof(float));
+	newint = calloc(newlen, sizeof(float));
 	
 	// printf("%d %d %d\n", oldlen, newlen, mzbins);
 	for (int i = 0; i < newlen; i++)
 	{
-		double mz = 0;
-		double val = 0;
-		double bins = 0;
+		float mz = 0;
+		float val = 0;
+		float bins = 0;
 		for(int j=0; j<mzbins; j++)
 		{
 			int index = i*mzbins + j;
@@ -51,11 +51,11 @@ int pool1d(double *oldmz, double *oldint, const int oldlen, const int mzbins)
 	}
 
 	//Bookkeeping 
-	realloc(oldmz, newlen * sizeof(double));
-	realloc(oldint, newlen * sizeof(double));
+	realloc(oldmz, newlen * sizeof(float));
+	realloc(oldint, newlen * sizeof(float));
 
-	memcpy(oldmz, newmz, sizeof(double)*newlen);
-	memcpy(oldint, newint, sizeof(double)*newlen);
+	memcpy(oldmz, newmz, sizeof(float)*newlen);
+	memcpy(oldint, newint, sizeof(float)*newlen);
 
 	free(newmz);
 	free(newint);
@@ -65,13 +65,13 @@ int pool1d(double *oldmz, double *oldint, const int oldlen, const int mzbins)
 
 
 //Chop data
-int chop1d(double *oldmz, double *oldint, int oldlen, const double min, const double max)
+int chop1d(float *oldmz, float *oldint, int oldlen, const float min, const float max)
 {
-	double *newmz, *newint;
+	float *newmz, *newint;
 	int newlen = 0;
 
-	double maxval = max1d(oldmz, oldlen);
-	double minval = min1d(oldmz, oldlen);
+	float maxval = max1d(oldmz, oldlen);
+	float minval = min1d(oldmz, oldlen);
 	if (maxval<max && minval>min)
 	{
 		return oldlen;
@@ -92,8 +92,8 @@ int chop1d(double *oldmz, double *oldint, int oldlen, const double min, const do
 	}
 	
 	//declare memory for new arrays
-	newmz = calloc(newlen, sizeof(double));
-	newint = calloc(newlen, sizeof(double));
+	newmz = calloc(newlen, sizeof(float));
+	newint = calloc(newlen, sizeof(float));
 	
 	//Fill new arrays
 	newlen = 0;
@@ -108,11 +108,11 @@ int chop1d(double *oldmz, double *oldint, int oldlen, const double min, const do
 	}
 	
 	//Bookkeeping 
-	realloc(oldmz, newlen * sizeof(double));
-	realloc(oldint, newlen * sizeof(double));
+	realloc(oldmz, newlen * sizeof(float));
+	realloc(oldint, newlen * sizeof(float));
 	
-	memcpy(oldmz, newmz, sizeof(double)*newlen);
-	memcpy(oldint, newint, sizeof(double)*newlen);
+	memcpy(oldmz, newmz, sizeof(float)*newlen);
+	memcpy(oldint, newint, sizeof(float)*newlen);
 	
 	free(newmz);
 	free(newint);
@@ -121,11 +121,11 @@ int chop1d(double *oldmz, double *oldint, int oldlen, const double min, const do
 }
 
 //Remove duplicates
-int remove_duplicates(double *oldmz, double *oldint, int oldlen)
+int remove_duplicates(float *oldmz, float *oldint, int oldlen)
 {
-	double *newmz, *newint;
+	float *newmz, *newint;
 	int newlen = 1;
-	double val = 0;
+	float val = 0;
 
 
 	//get new length
@@ -143,8 +143,8 @@ int remove_duplicates(double *oldmz, double *oldint, int oldlen)
 	}
 
 	//declare memory for new arrays
-	newmz = calloc(newlen, sizeof(double));
-	newint = calloc(newlen, sizeof(double));
+	newmz = calloc(newlen, sizeof(float));
+	newint = calloc(newlen, sizeof(float));
 
 	//Fill new arrays
 	newlen = 0;
@@ -167,11 +167,11 @@ int remove_duplicates(double *oldmz, double *oldint, int oldlen)
 	newlen++;
 
 	//Bookkeeping 
-	realloc(oldmz, newlen * sizeof(double));
-	realloc(oldint, newlen * sizeof(double));
+	realloc(oldmz, newlen * sizeof(float));
+	realloc(oldint, newlen * sizeof(float));
 
-	memcpy(oldmz, newmz, sizeof(double)*newlen);
-	memcpy(oldint, newint, sizeof(double)*newlen);
+	memcpy(oldmz, newmz, sizeof(float)*newlen);
+	memcpy(oldint, newint, sizeof(float)*newlen);
 
 	//free(newmz);
 	//free(newint);
@@ -180,9 +180,9 @@ int remove_duplicates(double *oldmz, double *oldint, int oldlen)
 }
 
 //Remove middle zeros
-int remove_middle_zeros(double *oldmz, double *oldint, int oldlen)
+int remove_middle_zeros(float *oldmz, float *oldint, int oldlen)
 {
-	double *newmz, *newint;
+	float *newmz, *newint;
 	int newlen = 2;
 
 	//get new length
@@ -200,8 +200,8 @@ int remove_middle_zeros(double *oldmz, double *oldint, int oldlen)
 	}
 
 	//declare memory for new arrays
-	newmz = calloc(newlen, sizeof(double));
-	newint = calloc(newlen, sizeof(double));
+	newmz = calloc(newlen, sizeof(float));
+	newint = calloc(newlen, sizeof(float));
 
 	//Fill new arrays
 	newlen = 0;
@@ -223,11 +223,11 @@ int remove_middle_zeros(double *oldmz, double *oldint, int oldlen)
 	newlen++;
 
 	//Bookkeeping 
-	realloc(oldmz, newlen * sizeof(double));
-	realloc(oldint, newlen * sizeof(double));
+	realloc(oldmz, newlen * sizeof(float));
+	realloc(oldint, newlen * sizeof(float));
 
-	memcpy(oldmz, newmz, sizeof(double)*newlen);
-	memcpy(oldint, newint, sizeof(double)*newlen);
+	memcpy(oldmz, newmz, sizeof(float)*newlen);
+	memcpy(oldint, newint, sizeof(float)*newlen);
 
 	//free(newmz);
 	//free(newint);
@@ -236,9 +236,9 @@ int remove_middle_zeros(double *oldmz, double *oldint, int oldlen)
 }
 
 //Normalize
-void norm1d(double *dataInt, const int lengthmz)
+void norm1d(float *dataInt, const int lengthmz)
 {
-	double maxval = max1d(dataInt, lengthmz);
+	float maxval = max1d(dataInt, lengthmz);
 	if(maxval>0){
 		for (int i = 0; i < lengthmz; i++)
 		{
@@ -247,10 +247,10 @@ void norm1d(double *dataInt, const int lengthmz)
 	}
 }
 
-void background_subtract(double *dataInt, const double bsub, const int lengthmz)
+void background_subtract(float *dataInt, const float bsub, const int lengthmz)
 {
-	double *background;
-	background = calloc(lengthmz, sizeof(double));
+	float *background;
+	background = calloc(lengthmz, sizeof(float));
 	// Get Local Mins
 	for (int i=0; i < lengthmz; i++)
 	{
@@ -258,7 +258,7 @@ void background_subtract(double *dataInt, const double bsub, const int lengthmz)
 		int end = i + abs(bsub);
 		if (start < 0) { start = 0; }
 		if (end > lengthmz) { end = lengthmz; }
-		double min = dataInt[start];
+		float min = dataInt[start];
 		for (int j = start+1; j < end; j++)
 		{
 			if (dataInt[j] < min) { min = dataInt[j]; }
@@ -267,8 +267,8 @@ void background_subtract(double *dataInt, const double bsub, const int lengthmz)
 	}
 	// Smooth Local Mins and Subtract
 	int threshold = abs(bsub) * 6;
-	double *gaussian;
-	gaussian = calloc(threshold, sizeof(double));
+	float *gaussian;
+	gaussian = calloc(threshold, sizeof(float));
 	for (int i = 0; i < threshold; i++)
 	{
 		gaussian[i] = ndis(i, abs(bsub) * 3, abs(bsub));
@@ -281,7 +281,7 @@ void background_subtract(double *dataInt, const double bsub, const int lengthmz)
 		int end = i + abs(bsub)*3;
 		if (start < 0) { start = 0; }
 		if (end > lengthmz) { end = lengthmz; }
-		double val = 0;
+		float val = 0;
 		for (int j = start; j < end; j++)
 		{
 			val += background[j] * gaussian[j - start];
@@ -296,20 +296,20 @@ void background_subtract(double *dataInt, const double bsub, const int lengthmz)
 // A comparator function used by qsort 
 inline int compare(const void* a, const void* b)
 {
-	return (*(double*)a - *(double*)b);
+	return (*(float*)a - *(float*)b);
 }
 
 //Cut out the lowest x percent of the data
-int data_reduction(double* oldmz, double* oldint, int oldlen, const double redper)
+int data_reduction(float* oldmz, float* oldint, int oldlen, const float redper)
 {
-	double* newmz, * newint, *sortint, cutoff;
+	float* newmz, * newint, *sortint, cutoff;
 	int newlen = 0;
 	
 	//Create the sort array and load it with intensities
-	sortint = calloc(oldlen, sizeof(double));
-	memcpy(sortint, oldint, sizeof(double) * oldlen);
+	sortint = calloc(oldlen, sizeof(float));
+	memcpy(sortint, oldint, sizeof(float) * oldlen);
 	//Sort it
-	qsort(sortint, oldlen, sizeof(double), compare);
+	qsort(sortint, oldlen, sizeof(float), compare);
 	
 	//Find the index redper percent of the way through the list
 	int index = (int)redper * oldlen / 100.;
@@ -334,8 +334,8 @@ int data_reduction(double* oldmz, double* oldint, int oldlen, const double redpe
 	if (newlen < 3) { printf("Warning: Aggressive Data Reduction!!! Lower the data reduction percentage\n\n"); }
 
 	//declare memory for new arrays
-	newmz = calloc(newlen, sizeof(double));
-	newint = calloc(newlen, sizeof(double));
+	newmz = calloc(newlen, sizeof(float));
+	newint = calloc(newlen, sizeof(float));
 
 	//Fill new arrays
 	newlen = 0;
@@ -350,11 +350,11 @@ int data_reduction(double* oldmz, double* oldint, int oldlen, const double redpe
 	}
 
 	//Bookkeeping 
-	realloc(oldmz, newlen * sizeof(double));
-	realloc(oldint, newlen * sizeof(double));
+	realloc(oldmz, newlen * sizeof(float));
+	realloc(oldint, newlen * sizeof(float));
 
-	memcpy(oldmz, newmz, sizeof(double) * newlen);
-	memcpy(oldint, newint, sizeof(double) * newlen);
+	memcpy(oldmz, newmz, sizeof(float) * newlen);
+	memcpy(oldint, newint, sizeof(float) * newlen);
 
 	free(newmz);
 	free(newint);
@@ -366,8 +366,8 @@ int data_reduction(double* oldmz, double* oldint, int oldlen, const double redpe
 void process_data(int argc, char *argv[], Config config)
 {	
 	hid_t file_id;
-	double *dataMZ = NULL;
-	double *dataInt = NULL;
+	float *dataMZ = NULL;
+	float *dataInt = NULL;
 	char dataset[1024];
 	char outdat[1024];
 	char strval[1024];
@@ -380,8 +380,8 @@ void process_data(int argc, char *argv[], Config config)
 	file_id = H5Fopen(argv[1], H5F_ACC_RDWR, H5P_DEFAULT);
 	strjoin(dataset, "/raw_data", outdat);
 	int lengthmz = mh5getfilelength(file_id, outdat);
-	dataMZ = calloc(lengthmz, sizeof(double));
-	dataInt = calloc(lengthmz, sizeof(double));
+	dataMZ = calloc(lengthmz, sizeof(float));
+	dataInt = calloc(lengthmz, sizeof(float));
 	mh5readfile2d(file_id, outdat, lengthmz, dataMZ, dataInt);
 	//printf("Length of Data: %d \n", lengthmz);
